requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
ccxt==1.59.85
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
cryptography==35.0.0
dateparser==1.1.0
idna==3.3
lxml==4.6.4
multidict==5.2.0
multitasking==0.0.9
numpy==1.21.3
pandas==1.3.4
pandas-ta==0.3.14b0
pkg_resources==0.0.0
pycares==4.1.2
pycparser==2.20
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.10.23
requests==2.26.0
schedule==1.1.0
six==1.16.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.0.2
ujson==4.2.0
urllib3==1.26.7
websockets==9.1
yarl==1.6.3
yfinance==0.1.64
